refactor: Name 클래스 명 변경, 패키지 구조 변경
