+++ src/main/java/blackjack/domain/player/Player.java
@@ -8,13 +8,13 @@ import java.util.Objects;
 
 public abstract class Player {
 
-    private final Name name;
+    private final GamblerName gamblerName;
     private final List<Card> cards;
     private static final int MAX_HAND_VALUE = 21;
     private static final int ACE_ADDITIONAL_VALUE = 10;
 
     public Player(final String name) {
-        this.name = new Name(name);
+        this.gamblerName = new GamblerName(name);
         cards = new ArrayList<>();
     }
 
@@ -41,7 +41,7 @@ public abstract class Player {
     }
 
     public String getName() {
-        return name.getName();
+        return gamblerName.name();
     }
 
     abstract public List<Card> getOpenedCards();
@@ -64,11 +64,11 @@ public abstract class Player {
         if (this == o) return true;
         if (o == null || getClass() != o.getClass()) return false;
         Player player = (Player) o;
-        return Objects.equals(name, player.name);
+        return Objects.equals(gamblerName, player.gamblerName);
     }
 
     @Override
     public int hashCode() {
-        return Objects.hash(name);
+        return Objects.hash(gamblerName);
     }
 }
